refactor(scheduled-tasks): remove sqs support
BREAKING CHANGE: sqs support has been removed as it was no longer maintained
and used version has been deprecated
BREAKING CHANGE: redis strategy is now default and the naming scheme of
all types and classes have been changed to match
BREAKING CHANGE: The options passed to `clientOptions.tasks` can be flattened
to remove the `new ScheduledTaskRedisStrategy` and just provide options directly.
