+++ packages/scheduled-tasks/package.json
@@ -17,46 +17,18 @@
 			"import": "./register.mjs",
 			"require": "./register.js",
 			"types": "./register.d.ts"
-		},
-		"./register-redis": {
-			"import": "./register-redis.mjs",
-			"require": "./register-redis.js",
-			"types": "./register-redis.d.ts"
-		},
-		"./register-sqs": {
-			"import": "./register-sqs.mjs",
-			"require": "./register-sqs.js",
-			"types": "./register-sqs.d.ts"
-		},
-		"./types-redis": {
-			"import": "./dist/types-redis.mjs",
-			"require": "./dist/types-redis.js",
-			"types": "./dist/types-redis.d.ts"
-		},
-		"./types-sqs": {
-			"import": "./dist/types-sqs.mjs",
-			"require": "./dist/types-sqs.js",
-			"types": "./dist/types-sqs.d.ts"
 		}
 	},
 	"sideEffects": [
 		"./dist/register.js",
-		"./dist/register-redis.js",
-		"./dist/register-sqs.js",
 		"./register.js",
-		"./register.mjs",
-		"./register-redis.js",
-		"./register-redis.mjs",
-		"./register-sqs.js",
-		"./register-sqs.mjs"
+		"./register.mjs"
 	],
 	"homepage": "https://github.com/sapphiredev/plugins/tree/main/packages/scheduled-tasks",
 	"scripts": {
 		"lint": "eslint src --ext ts --fix",
-		"build": "tsc -b src && yarn esm:register && yarn esm:register-redis && yarn esm:register-sqs && yarn esm:default",
+		"build": "tsc -b src && yarn esm:register && yarn esm:default",
 		"esm:register": "gen-esm-wrapper dist/register.js dist/register.mjs",
-		"esm:register-redis": "gen-esm-wrapper dist/register-redis.js dist/register-redis.mjs",
-		"esm:register-sqs": "gen-esm-wrapper dist/register-sqs.js dist/register-sqs.mjs",
 		"esm:default": "gen-esm-wrapper dist/index.js dist/index.mjs",
 		"docs": "typedoc-json-parser",
 		"prepack": "yarn build",
@@ -66,14 +38,12 @@
 	"dependencies": {
 		"@sapphire/stopwatch": "^1.5.0",
 		"@sapphire/utilities": "^3.11.1",
+		"bullmq": "^3.12.0",
 		"tslib": "^2.5.0"
 	},
 	"devDependencies": {
 		"@favware/cliff-jumper": "^2.0.0",
-		"bullmq": "^3.12.0",
 		"gen-esm-wrapper": "^1.1.3",
-		"sqs-consumer": "^5.8.0",
-		"sqs-producer": "^2.2.0",
 		"typedoc": "^0.24.6",
 		"typedoc-json-parser": "^7.4.0",
 		"typescript": "^5.0.4"
@@ -84,14 +54,9 @@
 		"directory": "packages/scheduled-tasks"
 	},
 	"files": [
-		"dist/**/*.js*",
-		"dist/**/*.mjs*",
-		"dist/**/*.d*",
-		"./register.*",
-		"./register-redis.*",
-		"./register-sqs.*",
-		"./types-redis.*",
-		"./types-sqs.*"
+		"dist/",
+		"!dist/.tsbuildinfo",
+		"./register.*"
 	],
 	"engines": {
 		"node": ">=v14.18.0",
