--- packages/scheduled-tasks/register-redis.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/register-redis';
