--- packages/scheduled-tasks/register-redis.mjs
@@ -1,1 +1,0 @@
-export * from './dist/register-redis.mjs';
