--- packages/scheduled-tasks/register-sqs.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/register-sqs';
