--- packages/scheduled-tasks/register-sqs.js
@@ -1,4 +1,0 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-const tslib = require('tslib');
-
-tslib.__exportStar(require('./dist/register-sqs.js'), exports);
