--- packages/scheduled-tasks/register-sqs.mjs
@@ -1,1 +1,0 @@
-export * from './dist/register-sqs.mjs';
