+++ packages/scheduled-tasks/src/index.ts
@@ -2,16 +2,13 @@ import 'tslib';
 
 import type { ScheduledTaskHandler } from './lib/ScheduledTaskHandler';
 import type { ScheduledTaskStore } from './lib/structures/ScheduledTaskStore';
-import type { ScheduledTasksOptions } from './lib/types/ScheduledTasksOptions';
+import type { ScheduledTaskHandlerOptions } from './lib/types/ScheduledTaskTypes';
 
 export * from './lib/ScheduledTaskHandler';
 export * from './lib/structures/ScheduledTask';
 export * from './lib/structures/ScheduledTaskStore';
-export * from './lib/types/ScheduledTaskBaseStrategy';
-export * from './lib/types/ScheduledTaskCreateRepeatedTask';
 export * from './lib/types/ScheduledTaskEvents';
-export * from './lib/types/ScheduledTasksOptions';
-export * from './lib/types/ScheduledTasksTaskOptions';
+export * from './lib/types/ScheduledTaskTypes';
 
 declare module '@sapphire/pieces' {
 	interface Container {
@@ -25,6 +22,6 @@ declare module '@sapphire/pieces' {
 
 declare module 'discord.js' {
 	export interface ClientOptions {
-		tasks?: ScheduledTasksOptions;
+		tasks?: ScheduledTaskHandlerOptions;
 	}
 }
