--- packages/scheduled-tasks/src/lib/types/ScheduledTaskCreateRepeatedTask.ts
@@ -1,6 +1,0 @@
-import type { ScheduledTasksTaskOptions } from './ScheduledTasksTaskOptions';
-
-export interface ScheduledTaskCreateRepeatedTask {
-	name: string;
-	options: ScheduledTasksTaskOptions;
-}
