--- packages/scheduled-tasks/src/lib/types/ScheduledTasksOptions.ts
@@ -1,5 +1,0 @@
-import type { ScheduledTaskBaseStrategy } from './ScheduledTaskBaseStrategy';
-
-export interface ScheduledTasksOptions {
-	strategy: ScheduledTaskBaseStrategy;
-}
