--- packages/scheduled-tasks/src/register-sqs.ts
@@ -1,4 +1,0 @@
-import './index';
-import './register';
-
-export * from './lib/strategies/ScheduledTaskSQSStrategy';
