--- packages/scheduled-tasks/src/types-redis.ts
@@ -1,9 +1,0 @@
-export type {
-	BullClient,
-	ScheduledTaskRedisListRepeatedReturnType,
-	ScheduledTaskRedisStrategy,
-	ScheduledTaskRedisStrategyJob,
-	ScheduledTaskRedisStrategyListOptions,
-	ScheduledTaskRedisStrategyListRepeatedOptions,
-	ScheduledTaskRedisStrategyOptions
-} from './lib/strategies/ScheduledTaskRedisStrategy';
