--- packages/scheduled-tasks/src/types-sqs.ts
@@ -1,1 +1,0 @@
-export type { ScheduledTaskSQSStrategy, ScheduledTaskSQSStrategyMessageBody, SQSClient } from './lib/strategies/ScheduledTaskSQSStrategy';
