--- packages/scheduled-tasks/types-redis.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/types-redis';
