--- packages/scheduled-tasks/types-redis.mjs
@@ -1,1 +1,0 @@
-export * from './dist/types-redis.mjs';
