--- packages/scheduled-tasks/types-sqs.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/types-sqs';
