--- packages/scheduled-tasks/types-sqs.js
@@ -1,4 +1,0 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-const tslib = require('tslib');
-
-tslib.__exportStar(require('./dist/types-sqs.js'), exports);
