--- packages/scheduled-tasks/types-sqs.mjs
@@ -1,1 +1,0 @@
-export * from './dist/types-sqs.mjs';
